requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.7
django-filter==25.1
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyMySQL==1.1.1
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.1
whitenoise==6.9.0
